directory refactor, fix css and add thumbnail.png
