+++ components/bl-stepper-component/src/components/stepper-line.js
@@ -1,0 +1,15 @@
+export const StepperLine = (props) => {
+  const {
+    stepIndex,
+    steps,
+    stepperLineClass,
+  } = props;
+
+  return (
+    <>
+      {stepIndex !== steps.length - 1 && (
+        <span className={stepperLineClass}></span>
+      )}
+    </>
+  );
+};
