+++ components/bl-stepper-component/src/use-active-class-name.js
